+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Downgrade `insn` to `0.2.1` to avoid classpath collisions with (many popular) libraries that use older `ASM` versions
 * Upgrade JMH to latest version `1.24`
 * Rename `:profilers` result key to the more accurate `:secondary`
+* Return `:statistics` data in all cases that JMH does
+* Provide `:histogram` data in `:statistics`
 * BUGFIX: warmup `:select`s should only be run during warmup
 * BUGFIX: emit all selected benchmarks to single class to enable JMH `:group` functionality
 * BUGFIX: improper single-width prim loads with `:instrument` option